+++ src/pages/FriendList/FriendListPage.jsx
@@ -1,10 +1,12 @@
 import React from "react";
 import styled from "styled-components";
 import ListForm from "../../components/FriendList/ListForm";
+import ListHeader from "../../components/FriendList/ListHeader";
 
 export default function FriendListPage() {
   return (
     <PageContainer>
+      <ListHeader />
       <ListForm />
     </PageContainer>
   );
@@ -15,5 +17,5 @@ const PageContainer = styled.div`
   display: flex;
   flex-direction: column;
   align-items: center;
-  justify-content: center;
+  height: 100vh;
 `;
