+++ src/pages/MyPage/MyPage.jsx
@@ -17,4 +17,5 @@ const PageContainer = styled.div`
   flex-direction: column;
   align-items: center;
   justify-content: center;
+  height: 100vh;
 `;
